add unit tsts for get all users
took mockito out of the pictur for testing controllers. just using a real h2 db with refreshed data every time is more predictable and less boilerplate-y. Going to have to find a way to make sure the unit tests still use h2 when I move the real thing to postgres